Extract comments package
