+++ src/main/java/com/code4ro/legalconsultation/comment/mapper/CommentMapper.java
@@ -1,8 +1,8 @@
-package com.code4ro.legalconsultation.converters;
+package com.code4ro.legalconsultation.comment.mapper;
 
-import com.code4ro.legalconsultation.model.dto.CommentDetailDto;
-import com.code4ro.legalconsultation.model.dto.CommentDto;
-import com.code4ro.legalconsultation.model.persistence.Comment;
+import com.code4ro.legalconsultation.comment.model.dto.CommentDetailDto;
+import com.code4ro.legalconsultation.comment.model.dto.CommentDto;
+import com.code4ro.legalconsultation.comment.model.persistence.Comment;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.Mappings;
