+++ src/main/java/com/code4ro/legalconsultation/comment/model/persistence/Comment.java
@@ -1,8 +1,11 @@
-package com.code4ro.legalconsultation.model.persistence;
+package com.code4ro.legalconsultation.comment.model.persistence;
 
 import lombok.Getter;
 import lombok.Setter;
 import org.springframework.data.annotation.LastModifiedDate;
+import com.code4ro.legalconsultation.model.persistence.BaseEntity;
+import com.code4ro.legalconsultation.model.persistence.ApplicationUser;
+import com.code4ro.legalconsultation.model.persistence.DocumentNode;
 
 import javax.persistence.*;
 import java.util.Date;
