+++ src/main/java/com/code4ro/legalconsultation/comment/model/persistence/CommentStatus.java
@@ -1,0 +1,6 @@
+package com.code4ro.legalconsultation.comment.model.persistence;
+
+public enum CommentStatus {
+    APPROVED,
+    REJECTED
+}
