+++ src/main/java/com/code4ro/legalconsultation/converters/NumberOfCommentsMapper.java
@@ -2,7 +2,7 @@ package com.code4ro.legalconsultation.converters;
 
 import com.code4ro.legalconsultation.model.dto.documentnode.DocumentNodeDto;
 import com.code4ro.legalconsultation.model.persistence.DocumentNode;
-import com.code4ro.legalconsultation.repository.CommentRepository;
+import com.code4ro.legalconsultation.comment.repository.CommentRepository;
 import org.mapstruct.AfterMapping;
 import org.mapstruct.Mapper;
 import org.mapstruct.MappingTarget;
