--- src/main/java/com/code4ro/legalconsultation/model/persistence/CommentStatus.java
@@ -1,6 +1,0 @@
-package com.code4ro.legalconsultation.model.persistence;
-
-public enum CommentStatus {
-    APPROVED,
-    REJECTED
-}
