+++ src/main/java/com/code4ro/legalconsultation/service/impl/VoteServiceImpl.java
@@ -1,12 +1,12 @@
 package com.code4ro.legalconsultation.service.impl;
 
+import com.code4ro.legalconsultation.comment.service.CommentService;
 import com.code4ro.legalconsultation.config.security.CurrentUserService;
 import com.code4ro.legalconsultation.model.dto.VoteDto;
 import com.code4ro.legalconsultation.model.persistence.ApplicationUser;
-import com.code4ro.legalconsultation.model.persistence.Comment;
+import com.code4ro.legalconsultation.comment.model.persistence.Comment;
 import com.code4ro.legalconsultation.model.persistence.Vote;
 import com.code4ro.legalconsultation.repository.VoteRepository;
-import com.code4ro.legalconsultation.service.api.CommentService;
 import com.code4ro.legalconsultation.service.api.VoteService;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
