+++ src/main/java/com/code4ro/legalconsultation/user/service/UserService.java
@@ -6,7 +6,7 @@ import com.code4ro.legalconsultation.model.persistence.Invitation;
 import com.code4ro.legalconsultation.user.model.persistence.User;
 import com.code4ro.legalconsultation.user.model.dto.UserDto;
 import com.code4ro.legalconsultation.user.model.persistence.UserRole;
-import com.code4ro.legalconsultation.repository.UserRepository;
+import com.code4ro.legalconsultation.user.repository.UserRepository;
 import com.code4ro.legalconsultation.service.api.InvitationService;
 import com.code4ro.legalconsultation.service.api.MailApi;
 import com.fasterxml.jackson.databind.ObjectReader;
