+++ src/test/java/com/code4ro/legalconsultation/common/controller/AbstractControllerIntegrationTest.java
@@ -6,7 +6,7 @@ import com.code4ro.legalconsultation.model.persistence.Invitation;
 import com.code4ro.legalconsultation.model.persistence.InvitationStatus;
 import com.code4ro.legalconsultation.user.model.persistence.User;
 import com.code4ro.legalconsultation.repository.InvitationRepository;
-import com.code4ro.legalconsultation.repository.UserRepository;
+import com.code4ro.legalconsultation.user.repository.UserRepository;
 import com.code4ro.legalconsultation.service.impl.ApplicationUserService;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.apache.commons.lang3.StringUtils;
