+++ src/test/java/com/code4ro/legalconsultation/service/CommentServiceTest.java
@@ -1,16 +1,19 @@
 package com.code4ro.legalconsultation.service;
 
+import com.code4ro.legalconsultation.comment.mapper.CommentMapper;
+import com.code4ro.legalconsultation.comment.model.dto.CommentDto;
+import com.code4ro.legalconsultation.comment.model.persistence.Comment;
+import com.code4ro.legalconsultation.comment.model.persistence.CommentStatus;
+import com.code4ro.legalconsultation.comment.service.impl.CommentServiceImpl;
 import com.code4ro.legalconsultation.common.exceptions.LegalValidationException;
 import com.code4ro.legalconsultation.config.security.CurrentUserService;
-import com.code4ro.legalconsultation.converters.CommentMapper;
-import com.code4ro.legalconsultation.model.dto.CommentDto;
-import com.code4ro.legalconsultation.model.persistence.*;
-import com.code4ro.legalconsultation.repository.CommentRepository;
-import com.code4ro.legalconsultation.service.api.DocumentNodeService;
-import com.code4ro.legalconsultation.service.impl.CommentServiceImpl;
 import com.code4ro.legalconsultation.factory.CommentFactory;
 import com.code4ro.legalconsultation.factory.DocumentNodeFactory;
 import com.code4ro.legalconsultation.factory.RandomObjectFiller;
+import com.code4ro.legalconsultation.model.persistence.ApplicationUser;
+import com.code4ro.legalconsultation.model.persistence.DocumentNode;
+import com.code4ro.legalconsultation.comment.repository.CommentRepository;
+import com.code4ro.legalconsultation.service.api.DocumentNodeService;
 import com.code4ro.legalconsultation.user.model.persistence.UserRole;
 import org.junit.Before;
 import org.junit.Test;
@@ -37,6 +40,8 @@ import static org.mockito.Mockito.*;
 public class CommentServiceTest {
 
     public static final CommentStatus NEW_STATUS = CommentStatus.APPROVED;
+    private final DocumentNodeFactory documentNodeFactory = new DocumentNodeFactory();
+    private final CommentFactory commentFactory = new CommentFactory();
     @Mock
     private CommentRepository commentRepository;
     @Mock
@@ -47,12 +52,8 @@ public class CommentServiceTest {
     private DocumentNodeService documentNodeService;
     @InjectMocks
     private CommentServiceImpl commentService;
-
     @Captor
     private ArgumentCaptor<Comment> commentArgumentCaptor;
-
-    private final DocumentNodeFactory documentNodeFactory = new DocumentNodeFactory();
-    private final CommentFactory commentFactory = new CommentFactory();
     private ApplicationUser currentUser;
 
     @Before
@@ -144,7 +145,7 @@ public class CommentServiceTest {
     }
 
     @Test
-    public  void updateCommentStatus(){
+    public void updateCommentStatus() {
         final UUID id = UUID.randomUUID();
         final Comment comment = new Comment();
         comment.setStatus(null);
@@ -155,7 +156,7 @@ public class CommentServiceTest {
     }
 
     @Test(expected = RuntimeException.class)
-    public  void failUpdateCommentStatus(){
+    public void failUpdateCommentStatus() {
         final UUID id = UUID.randomUUID();
         final Comment comment = new Comment();
         comment.setStatus(CommentStatus.REJECTED);
@@ -164,7 +165,7 @@ public class CommentServiceTest {
     }
 
     @Test(expected = EntityNotFoundException.class)
-    public  void attemptToUpdateNonExistingComment(){
+    public void attemptToUpdateNonExistingComment() {
         final UUID id = UUID.randomUUID();
         final Comment comment = new Comment();
         comment.setStatus(CommentStatus.REJECTED);
