+++ src/test/java/com/code4ro/legalconsultation/service/DocumentServiceTest.java
@@ -1,21 +1,24 @@
 package com.code4ro.legalconsultation.service;
 
+import com.code4ro.legalconsultation.comment.service.CommentService;
 import com.code4ro.legalconsultation.converters.DocumentConsolidatedMapper;
 import com.code4ro.legalconsultation.factory.RandomObjectFiller;
 import com.code4ro.legalconsultation.model.dto.DocumentConsolidatedDto;
 import com.code4ro.legalconsultation.model.dto.DocumentMetadataDto;
 import com.code4ro.legalconsultation.model.persistence.DocumentConsolidated;
-import com.code4ro.legalconsultation.user.model.persistence.User;
+import com.code4ro.legalconsultation.model.persistence.DocumentMetadata;
+import com.code4ro.legalconsultation.model.persistence.DocumentExportFormat;
+import com.code4ro.legalconsultation.model.persistence.DocumentNode;
+import com.code4ro.legalconsultation.model.persistence.PdfHandle;
 import com.code4ro.legalconsultation.service.api.MailApi;
-import com.code4ro.legalconsultation.model.persistence.*;
-import com.code4ro.legalconsultation.service.api.CommentService;
 import com.code4ro.legalconsultation.service.export.DocumentExporterFactory;
 import com.code4ro.legalconsultation.service.export.PDFExporter;
 import com.code4ro.legalconsultation.service.impl.DocumentConsolidatedService;
 import com.code4ro.legalconsultation.service.impl.DocumentMetadataService;
 import com.code4ro.legalconsultation.service.impl.DocumentServiceImpl;
-import com.code4ro.legalconsultation.user.service.UserService;
 import com.code4ro.legalconsultation.service.impl.pdf.PDFServiceImpl;
+import com.code4ro.legalconsultation.user.model.persistence.User;
+import com.code4ro.legalconsultation.user.service.UserService;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.ArgumentCaptor;
